myP KeyError issue is solved.
The wait_for_robot function of the prob_interface returned immediately after call because the robot status has not been checked.
Since the wait_for_robot function did not wait as expected for the completion of the initialization, a call of the
calibration function directly after initialization raised an error in myP. Now the wait function works as expected.